gateway/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
grpcio==1.75.0
protobuf>=4.25.0
httpx==0.27.2
prometheus-client==0.21.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
uvloop==0.20.0
aiofiles==24.1.0

